Only run libwbclient %pre on upgrade, not fresh install
This seems to be breaking live image build tests because the
script is running before `rm` is installed. I do not know why
it didn't fail the tests run *on the update itself*, but it's
now causing tests of subsequent updates to fail, which is a
problem. Probably an ordering issue.

AFAIK, this script is meant to clean up stuff from earlier
versions of the package, so it's not relevant on fresh installs,
and we can just skip it and avoid any problems in the fresh
install case.